[performance] make rule evaluations side output into trace logger
In the current state, this stream is only needed for debugging / showing every
action in the console which can be better achieved by enabling a logger. If the
transaction throughput increases, this would have been a bottleneck.

log4j2.properties speficies a new logger for getting the output back in a
similar fashion with added thread info.